+++ packages/features/bookings/types.ts
@@ -1,7 +1,7 @@
 import type { ErrorOption, FieldPath } from "react-hook-form";
 
 import type { RegularBookingCreateResult } from "@calcom/features/bookings/lib/dto/types";
-import type { Slots } from "@calcom/features/calendars/lib/slots";
+import type { Slots } from "@calcom/features/calendars/lib/types";
 import type { SchedulingType } from "@calcom/prisma/enums";
 import type { RouterOutputs } from "@calcom/trpc/react";
 import type { AppsStatus } from "@calcom/types/Calendar";
